backend/requirements.txt
# requirements.txt (修正版)

Flask==3.0.3
Flask-Cors==4.0.1
feedparser==6.0.11
requests==2.32.3
google-generativeai==0.7.1
gunicorn==22.0.0
google-cloud-secret-manager==2.20.0  # <-- 【重要】不足していたライブラリを追加